gns3-py/requirements.txt
gns3fy
tabulate
python-dotenv
pathlib